Reviewed by Maciej Stachowiak, Dave Hyatt.
        Changed the dead resource LRU-SP algorithm to measure an object's
        total size, not just its encoded size. This will allow us to make
        better decisions about what data to evict when the cache is small. For
        example, the PLT can now run with a 16MB cache without fully evicting
        any resources.

        (Previously, we had assumed that decoded size would be an OK estimate
        of encoded size, but that is not true of GIF, whose decoded size can be
        orders of magnitude greater than its encoded size.)

        Subtly, destroying a resource's decoded data now increases its recency
        by moving it to the head of a smaller LRU list. This is slightly odd,
        but, since all resources get the same treatment, it shouldn't hurt
        the eviction algorithm.

        * history/PageCache.cpp:
        (WebCore::PageCache::releaseAutoreleasedPagesNow): Make sure that a
        dead resource eviction doesn't happen until we've released all of our
        dead pages. Otherwise, the cache will make terrible decisions about
        what to evict because all of our dead resources will seem live.

        * loader/Cache.cpp:
        (WebCore::Cache::Cache):
        (WebCore::Cache::pruneLiveResources):
        (WebCore::Cache::pruneDeadResources): Removed call to
        removeFromLiveDecodedResourcesList because this happens automatically
        now as a part of the process of changing the resource's decoded size.
        (WebCore::Cache::lruListFor): *** The key change. *** Compute the
        appropriate LRU list based on total size, not encoded size.
        (WebCore::Cache::dumpLRULists): Added debug logging function to help
        visualize the cache.

        * loader/Cache.h:
        (WebCore::Cache::setDeadResourcePruneEnabled):
        (WebCore::Cache::deadResourcePruneEnabled):

        * loader/CachedImage.cpp: Moved decoded size tracking code from here
        up into the base class. Currently, only CachedImage has a use for that
        functionality, but other subclasses might need it in the future, and
        the base class is already responsible for similar code related to
        encoded size tracking.
        (WebCore::CachedImage::decodedSizeChanged):
        * loader/CachedImage.h:

        * loader/CachedResource.cpp:
        (WebCore::CachedResource::CachedResource):
        (WebCore::CachedResource::setDecodedSize): Move us in the LRU-SP list
        just like setEncodedSize does, since decoded size counts now, too.
        (WebCore::CachedResource::setEncodedSize): Changed slightly to match
        the style of setDecodedSize.

        * loader/CachedResource.h:
        (WebCore::CachedResource::decodedSize):



Canonical link: https://commits.webkit.org/19576@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@25025 268f45cc-cd09-0410-ab3c-d52691b4dbfc

+++ WebCore/loader/Cache.cpp
@@ -50,6 +50,7 @@ Cache* cache()
 
 Cache::Cache()
 : m_disabled(false)
+, m_deadResourcePruneEnabled(true)
 , m_maximumSize(cDefaultCacheSize)
 , m_currentSize(0)
 , m_liveResourcesSize(0)
@@ -170,17 +171,16 @@ void Cache::pruneLiveResources()
     while (current) {
         CachedResource* prev = current->m_prevInLiveResourcesList;
         ASSERT(current->referenced());
-        if (current->isLoaded()) {
+        if (current->isLoaded() && current->decodedSize()) {
             // Check to see if the remaining resources are too new to prune.
             double elapsedTime = currentTime - current->m_lastDecodedAccessTime;
             if (elapsedTime < cMinDelayBeforeLiveDecodedPrune)
                 return;
 
-            // Go ahead and destroy our decoded data.
+            // Destroy our decoded data. This will remove us from 
+            // m_liveDecodedResources, and possibly move us to a differnt LRU 
+            // list in m_allResources.
             current->destroyDecodedData();
-            
-            // Now remove us from the list so the next prune can ignore us.
-            removeFromLiveDecodedResourcesList(current);
             
             // Stop pruning if our total live resource size is back under the maximum.
             if (m_liveDecodedSize <= m_maximumSize)
@@ -204,7 +204,10 @@ void Cache::pruneDeadResources()
     if (unreferencedResourcesSize < m_maximumSize)
         return;
     
-    unsigned size = m_allResources.size();
+    if (!m_deadResourcePruneEnabled)
+        return;
+        
+    int size = m_allResources.size();
     bool canShrinkLRULists = true;
     for (int i = size - 1; i >= 0; i--) {
         // Remove from the tail, since this is the least frequently accessed of the objects.
@@ -213,8 +216,10 @@ void Cache::pruneDeadResources()
         // First flush all the decoded data in this queue.
         while (current) {
             CachedResource* prev = current->m_prevInAllResourcesList;
-            if (!current->referenced() && current->isLoaded()) {
-                // Go ahead and destroy our decoded data.
+            if (!current->referenced() && current->isLoaded() && current->decodedSize()) {
+                // Destroy our decoded data. This will remove us from 
+                // m_liveDecodedResources, and possibly move us to a differnt 
+                // LRU list in m_allResources.
                 current->destroyDecodedData();
                 
                 // Stop pruning if our total cache size is back under the maximum or if every
@@ -317,7 +322,7 @@ static inline unsigned fastLog2(unsigned i)
 Cache::LRUList* Cache::lruListFor(CachedResource* resource)
 {
     unsigned accessCount = max(resource->accessCount(), 1U);
-    unsigned queueIndex = fastLog2(resource->encodedSize() / accessCount);
+    unsigned queueIndex = fastLog2(resource->size() / accessCount);
 #ifndef NDEBUG
     resource->m_lruIndex = queueIndex;
 #endif
@@ -575,4 +580,23 @@ void Cache::setDisabled(bool disabled)
     }
 }
 
+#ifndef NDEBUG
+void Cache::dumpLRULists(bool includeLive) const
+{
+    printf("LRU-SP lists in eviction order (Kilobytes decoded, Kilobytes encoded, Access count, Referenced):\n");
+
+    int size = m_allResources.size();
+    for (int i = size - 1; i >= 0; i--) {
+        printf("\n\nList %d: ", i);
+        CachedResource* current = m_allResources[i].m_tail;
+        while (current) {
+            CachedResource* prev = current->m_prevInAllResourcesList;
+            if (includeLive || !current->referenced())
+                printf("(%.1fK, %.1fK, %uA, %dR); ", current->decodedSize() / 1024.0f, current->encodedSize() / 1024.0f, current->accessCount(), current->referenced());
+            current = prev;
+        }
+    }
 }
+#endif
+
+} // namespace WebCore
